139979 by tomj on 2019/07/10 (pair: msakai) UDSUP-8 Merge change 139905 to Jasper 139905 by tomj on 2019/07/08 (pair: msakai,sweettea) UDSUP-8 use one name for the chunkName/hash field of a UDS request
The public name of the UdsRequest field containing the hash value is
"chunkName".  The name of the same field in the private internal
Request is "hash".  Make them use the same name.

.../request.h
.../index.c
.../indexRouter.c
.../indexZone.c
.../volume.c
Rename the "hash" field of Request to use the same name used by a
UdsRequest, which is "chunkName".

139985 by corwin on 2019/07/10 (pair: sclafani)
Merged 138779 to linux-vdo.
138779 by corwin on 2019/04/12 (pair: msakai)
Converted slab depot save to not require SlabCompletion [VDO-4402].

This is the next step in unifying the drain operations on the slab
depot and also begins to implement a true suspend in the depot which
quiesces without writing dirty slab journal or reference count blocks.

.../blockAllocator.h
.../blockAllocatorInternals.h
.../blockAllocator.c
Renamed closeBlockAllocator() to drainBlockAllocator() also renaming
related fields and enumerants.
Created a SlabActor sub-structure to hold the fields needed for
iterating over slabs and factored out generic machinery for doing
so from loadBlockAllocator().
Renamed doCloseStep() to doDrainStep() and converted it to use the
applyToSlabs() method to call drainSlab() on the slabs instead of
using the SlabCompletion to save the slabs.
Removed unused 'saveRequested' field.

.../slabJournal.h
.../slabJournal.c
Removed closeSlabJournal() moving the only thing it did (an assertion)
before calling oldDrainSlabJournal() into drainSlabJournal().

.../refCounts.h
.../refCountsInternals.h
.../refCounts.c
Removed closeReferenceCounts() and added support for
ADMIN_STATE_SAVING and ADMIN_STATE_SUSPENDING to drainRefCounts().
Removed the 'closeRequested' field and updated checks against it to
check the RefCounts' AdminState instead.
Modified checkForDrainComplete() to consider itself complete when
suspending even if there are dirty blocks. Replaced
isRefCountsDirty() with isRefCountsActive() to capture this
difference.
Updated finishReferenceBlockWrite(), the only other caller of
isRefCountsDirty() to use isRefCountsActive() and an explicit dirty
check.

.../slabCompletion.h
.../slabCompletion.c
Removed saveSlabs() which was used by closeBlockAllocator().

.../slabDepot.c
Reflected change to BlockAllocator.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>