140009 by tomj on 2019/07/11 (pair: sclafani) UDSUP-7 Merge change 139913 to Jasper 139913 by tomj on 2019/07/08 (pair: msakai) UDSUP-7 Index replay checks the entire volume in a block device.
The code to find the index chapters to replay starts by finding the
valid chapters in the volume.  When we write the volume as a filesystem
file, this started by asking the size of the file.  This is now
embodied in the call to the getRegionDataSize method.

On a block device, this getRegionDataSize method returns the size
of the volume allocated within the block device.  And in user mode,
we now have a single file that acts like the block device, so the
getRegionDataSize method will return the size of the volume.

So it is time to stop using the getRegionDataSize method and just use
the size of the volume.

.../volume.c
In the findVolumeChapterBoundaries method, dispense with the call to
the getRegionDataSize method and the associated calculations.  Just
use the size of the volume region from the index geometry.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>