140018 by tomj on 2019/07/11 (pair: msakai) Merge change 139922 to Jasper 139922 by tomj on 2019/07/09 (pair: sweettea) Clean up user mode process/thread code.
Delete the running of our own shutdown code at program exit time.
This code was built under control of the DESTRUCTOR macro, and was
only used in user mode.  We no longer need to run our shutdown code.

Delete the udsSetShutdownHook method.  We used it to modify the
shutdown sequence so that we could do other shutdown actions either
before or after the standard shutdown, but under our control as to the
order of the actions.  Since we no longer automatically save any open
indices, we do not need this method.

Delete the udsSetThreadStartHook method.  We created this method for
the Oracle release.  Other than to test that it works, we have never
used it.  It may or may not be a good idea, but it exists only in user
mode and we do not use it.

.../udsState.c
Do not use a destructor to run our own shutdown code.

.../uds.h
.../udsState.c
Delete the udsSetShutdownHook method.

.../uds.h
.../threadsLinuxUser.c
Delete the udsSetThreadStartHook method.

.../userLinux/uds/featureDefs.h
Delete the DESTRUCTOR feature macro.  It is no longer used.

140019 by tomj on 2019/07/11 (pair: msakai)
Merge changes 139912, 139937 and 139938 to Jasper
139912 by tomj on 2019/07/08 (pair: msakai)
Delete the getComponentSizeForPortal method.

.../indexComponent.h
.../indexComponent.c
Delete the unused getComponentSizeForPortal method.

139938 by tomj on 2019/07/09 (pair: msakai)
Delete the histogram module.

.../histogram.c
.../histogram.h
We do not use this module.  I copied the module from VDO with plans
to be using it, but other things came up.  If we really want to do
histograms in UDS, we should copy a more recent version from
linux-vdo which is more prepared to be upstreamed.

140030 by tomj on 2019/07/12 (pair: sclafani)
Merge change 139951 to Jasper
139951 by tomj on 2019/07/09 (pair: msakai)
Eliminate the duplicate copy of the index session from a request

.../request.h
.../request.c
.../indexSession.c
.../volume.c
The "session" and "indexSession" fields of a Request contain exactly
the same value.  Eliminate the indexSession field and just use the
session field.

140033 by tomj on 2019/07/12 (pair: msakai)
Merge change 139939 to Jasper.
139939 by tomj on 2019/07/09 (pair: msakai)
Use static initialization for the error message table

.../errors.c
Use a static initializer for the registeredErrors data structure.
Remove the code that works very hard to initialze that structure
exactly once.  This includes the initializeStandardErrorBlocks and
ensureStandardErrorBlocks methods.

.../errors.h
.../errors.c
.../udsState.c
Remove all the calls to the ensureStandardErrorBlocks method.

140041 by corwin on 2019/07/12 (pair: msakai)
Integrate 138939 to linux-vdo.
138939 by corwin on 2019/04/25 (pair: msakai)
Convert SlabSummary to support suspend/resume [VDO-4402].

This change converts the slab summary to support suspend and resume
and completes the implementation of suspend on the slab depot.

.../slabSummary.h
.../slabSummaryInternals.h
.../slabSummary.c
Combined suspendSlabSummaryZone(), saveSlabSummaryZone(), and
closeSlabSummaryZone() into drainSlabSummaryZone(). In so doing,
replaced the 'saveWaiter', 'pendingAction', and 'suspended' fields of
the SlabSummaryZone with a AdminState.
Modified resumeSlabSummaryZone() to wait for the write of the summary
to complete so that a future resume transaction can get an
appropriate error and fail the resume if writing the new summary
fails.
Renamed checkForSaveComplete() to checkForOperationComplete() and
modified it to use the AdminState since it is now used to check for
completeness of both drains and resumes of summary zones.
Modified updateSlabSummaryEntry() to not allow updates if the summary
is quiescent or draining.
Modified loadSlabSummary() to use the new AdminState in SlabSummaryZone
0. Removed the 'loadParent' field.
Modified loadSlabSummary() to set finishCombiningZones() as the error
handler for reading the summary and removed duplicate error handling
code from from finishLoadingSummary() which is still the success
callback.

.../blockAllocator.h
.../blockAllocator.c
Modified doDrainStep() to call drainSlabSummaryZone() with the
allocator's current drain operation instead of calling
closeSlabSummaryZone(). As a result, drainBlockAllocator() is now
used for suspend as well as save (nee close).
Removed suspendSummaryZone() now that drainBlockAllocator() does
suspend.
Converted finishFlushingSlabJournals() to call drainSlabSummary().
Added resumeBlockAllocator() and doResumeStep() which goes through the
same set of steps as drainBlockAllocator() but in reverse (resume
the summary, then the slabs, and finally the scrubber). This is
necessary now that suspend applies to the entire allocator instead of
just the summary.
Removed resumeSummaryZone().
Modified applyToSlabs() to clear the allocator's open slab since the
ensuing requeue which now happens on resume will render the old
open slab invalid.
Modified queueSlab() to handle being called when resuming in addition
to when slabs are first being added to the allocator:
If the slab is unrecovered, register it for scrubbing.
Otherwise, prioritize it as normal, but don't update the statistics
which were already updated the first time the slab was queued.

.../slabDepot.c
Modified drainSlabDepot() to use drainBlockAllocator() for suspend.
Modified resumeSlabDepot() to use resumeBlockAllocator().

.../adminState.h
.../adminState.c
Fixed bug in endOperation() which would set an incorrect next state for
quiescent operations due to the fact that the next state of a
quiescent operation depends upon the state prior to the operation.
However, by the time endOperation() is called, what that prior state
was is no longer known. The bug is fixed by introducing a
'nextState' field to AdminState which is set in beginOperation() when
the determination can be made. Added isQuiescingCode() for use
in making the next state determination.
Added predicates, assert, start, and finish methods for resuming
analogous to the drain methods. Added finishOperationWithResult() for
use by slabSummary's checkForOperationComplete().

.../slab.h
.../slab.c
Added isSlabResuming().
Modified resumeSlab() to be callable via applyToSlabs() by giving it
the same signature as drainSlab().
Modified resumeSlab() to queue the Slab since applyToSlabs() removes
slabs from any ring they may be on.

.../slabScrubber.h
.../slabScrubber.c
Modified resumeScrubbing() to take a parent completion so that it has
the same form as other resume steps.
Fixed bug in resumeScrubbing() by not actually attempting to resume and
scrub a slab if there are no slabs to scrub.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>