140052 by tomj on 2019/07/15 (pair: msakai) UDSUP-7 Merge change 140024 to Jasper 140024 by tomj on 2019/07/11 (pair: msakai) UDSUP-7 Keep the IOFactory in the IndexLayout.
We have been using the IOFactory to make us an IORegion for the entire
index and keeping that IORegion in the IndexLayout.  In order to use
dm-bufio, we plan for the IOFactory to make us a dm_bufio_client.
To prepare for this, we keep a pointer to the IOFactory in the
IndexLayout.

This did not turn out to be easy and straightforward, so this change
refactors the IndexLayout code to make the handling of the IOFactory
and IORegion straightforward, merges some common code paths, and
simplifies the freeing of the IndexLayout.

.../indexLayout.c
Refactor some common code into the makeIndexLayoutFromFactory method.
It is where we allocate the IndexLayout, and free it if there were
any errors initializing it.  It is also where we set the factory,
region and refCount fields.
Unify all the destructor code paths, so that no calls remain to
destroySingleFileLayout.  Everything that wants to destroy the
IndexLayout calls freeIndexLayout.  The  destroySingleFileLayout
method has been refactored into the freeIndexLayout method.
Eliminate the DEFAULT_BLOCK_SIZE constant.  Instead we use the
UDS_BLOCK_SIZE constant.  It is important that all of UDS uses the
same 4K block size, so we should have only 1 named constant.
Eliminate the validateOffsetAndBlockSize method.  The validation
performed by this method are also done when the IOFactory makes an
IORegion, and in a way that is appropriate for the device or file
being used.
Rename the useCount field to refCount, which seems to be the more
common name for a reference counter.
Delete the loaded and saved flags, which are set but never used.
Delete the close flag, which was used to sometimes disable the closing
of the IORegion.  This was hard to follow, but we have now eliminated
the need for the flag.

140056 by corwin on 2019/07/15 (pair: sclafani)
Merged 139082 to linux-vdo.
139082 by corwin on 2019/05/07 (pair: msakai, sweettea, sclafani)
Use physical zone threads when recovering slab journals [VDO-4587].

If, during recovery, more than one slab journal block's worth of
entries are played into any single slab journal, that journal will need
to commit its tail block. The replay and commit are launched from the
logical zone thread 0. Unfortunately, the VIO used for doing the commit
is set to call back on the physical zone thread of the slab journal. If
a second commit happens, two threads could end up modifying the
journal's VIO pool without appropriate memory fencing. Furthermore, the
journal itself and also the slab summary may also be so affected.

This change reworks the recovery transaction to access the slab
journals on their correct threads by replaying journal entries one zone
at a time. It leverages the new action manager machinery in the slab
depot to do so.

This change also introduces thread error checking to VIOPool.

.../vdoRecovery.h
.../vdoRecoveryInternals.h
.../vdoRecovery.c
Reworked the recovery process to only access slab journals from their
physical zone threads. Recovery is now reordered:
- First, the recovery journal is read and analyzed on the journal
thread.
- Next, on logical zone 0, lookups are done to fill in any missing
decrefs which can not be resolved with the journal alone.
- Next, on the admin thread, the slab depot is loaded. The depot load
already loads each block allocator in turn, with each load
occurring on that allocator's zone thread. As part of the allocator
load, after the slab summary and slab journal tails have been read,
the allocator recovers its slab journals. This is done by calling
the new replayIntoSlabJournals() method. This method replays all
journal entries and synthsized decrefs for the calling allocator's
zone.
- Once the depot is loaded and recovered, the recovery continues as
before by draining the slab depot, saving the super block, and then
recovering the block map.
If missing decrefs are replayed more than once (due to a crash during
recovery), it is necessary that the fake journal points for them
are consistent for a given recovery journal. To support this,
MissingDecrefs were converted to be Waiters instead of RingNodes.
As each MissingDecref is created, it is assigned its journal point
and placed on a queue.
Modified RecoveryCompletion to have an array of WaitQueues for holding
MissingDecrefs which are ready to be applied. There is one queue for
each physical zone. Once the penultimate mappings for all
MissingDecrefs are known, they are placed on the queues for the
physical zones to which they apply. The remaining fields used to
manage MissingDecrefs were replaced with two new fields:
'incompleteDecrefsCount' and 'nextSynthesizedJournalPoint'.
Added an 'allocator' field to hold the BlockAllocator which is
currently being recovered.
Made addSlabJournalEntries() static since it had no external callers.

.../blockAllocator.h
.../blockAllocator.c
Modified finishLoadingAllocator() to recover the allocator's slab
journals if the load is for recovery.
Added notifySlabJournalsAreRecovered() to be called from vdoRecovery
once an allocator's slab journals have been recovered.

.../slabDepot.h
.../slabDepot.c
Modified loadSlabDepot() to take an optional 'context' parameter which
is passed to scheduleOperationWithContext(). This is used to store
the recovery completion so that each block allocator can retrieve it
and rebuild its slabs when loading for recovery.

.../actionManager.h
.../actionManager.c
Added scheduleOperationWithContext() and added a context field to
Action which can be retrieved with the new getCurrentActionContext().
This is used to get the recovery completion out of the action manager
so the block allocator can rebuild its slabs during recovery.

.../readOnlyRebuild.c
.../vdoLoad.c
Reflected change to loadSlabDepot().

.../vioPool.h
.../vioPool.c
Added a 'threadID' field and constructor parameter to track which
thread each VIOPool is to be used on.
Modified acquireVIOFromPool() and returnVIOToPool() to assert that they
are called on the correct thread. This error checking was used both
to demonstrate this bug and to validate the fix.

.../vdoInternal.h
.../vdo.c
Added assertOnPhysicalZoneThread().

.../types.h
Added ZONE_TYPE_ADMIN.

.../blockAllocator.c
.../blockMapTree.c
Reflected change to makeVIOPool().

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>