140064 by corwin on 2019/07/15 (pair: msakai) Merged 139091 to linux-vdo. 139091 by corwin on 2019/05/08 (pair: msakai) Removed use of slab journal completions for recovery [VDO-4402].
As a precursor to the final step in removing SlabCompletion, this
change removes the use of the slab journal's completion when waiting
for slab journal space during recovery. When combined with the removal
of the SlabCompletion mechanisms, this will allow the slab journal
completion to go away.

.../slabJournal.h
.../slabJournalInternals.h
.../slabJournal.c
Combined mayAddSlabJournalEntry() and addSlabJournalEntryForRebuild()
into a single new method, attemptReplayIntoSlabJournal(). The new
method uses the journal's slab's AdminState to hold the parent
completion instead of using the slab journal's completion.
Removed the 'waitingForSpace' field in lieu of the AdminState.

.../adminState.h
.../adminState.c
Added ADMIN_STATE_WAITING_FOR_RECOVERY for use by slab journals. This
state replaces the journal's 'waitingForSpace' field.
Added startOperationWithWaiter().
Added assertOperation() and factored common code out of all the
assert*Operation() methods into a new checkCode() method.

.../vdoRecovery.c
Reflected changes to slab journal.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>