140385 by corwin on 2019/08/15 (pair: msakai) Merged 139590 to linux-vdo [BZ 1659303]. 139590 by corwin on 2019/06/11 (pair: msakai) Fixed [VDO-4402]: do not write while suspended.
This is the final change required to fix VDO-4402. There will likely be
some follow-on clean-ups, but they are not necessary to fix the bug.

This change introduces a resume admin operation. It rewrites both growth
operations to only be valid when the VDO is suspended and to not do any
suspending or resuming of components themselves. In addition given that
the growth operations function during the dm pre-resume phase, from
which an error only results in the device not being resumed, reverting
to the pre-growth state on a failed growth is no longer done.

.../vdoResume.h
.../vdoResume.c
Created a resume admin operation.

.../adminCompletion.h
.../adminCompletion.c
.../vdoSuspend.c
Added ThreadIDGetterForPhase() type and added a 'getThreadID' field to
AdminCompletion in order to make the machinery for setting and
checking thread IDs based upon the operation phase common for all
admin operations. Added a 'threadIDGetter' parameter to
performAdminOperation().
Exposed assertAdminOperationType().
Added assertAdminPhaseThread().
Factored resetSubTask() into adminCompletion from vdoSuspend and
renamed it to resetAdminSubTask().

.../vdoSuspend.c
Modified SUSPEND_PHASE_PACKER to check for read-only mode in order to
ensure that the suspend operation will get a read-only mode error and
not an invalid admin state error in the case where a prior resume did
not resume all the components due to the VDO already being read-only.

.../adminState.h
.../adminState.c
Added ADMIN_STATE_NEW, a new quiescent state for use by grown slabs
which are quiescent when first created, but not suspended or saved.
Modified beginOperation() to reject quiescent operations when the VDO
is not quiescent. Now that we have the full suspend and resume
semantics, we can enfore quiescent operations only being allowed from
quiescent states.

.../vdoResize.c
.../vdoResizeLogical.c
Rewrote to use the new phase machinery in AdminCompletion.
Removed the reversion to the old state on error.
Removed all suspend and resume steps since these operations are now
only allowed on suspended VDO's and the resume will be handled after
they are done.

.../blockAllocator.h
.../blockAllocator.c
Removed the 'resizing' parameter from registerSlabWithAllocator(). This
parameter was used to indicate that the slab being registered should
be prioritized immediately. Since the allocator is now suspended
during grow physical, and all the slabs will be prioritized on
resume, we can not and should not prioritize the new slabs here.
Fixed doc for queueSlab().

.../blockMap.h
.../blockMap.c
Modified growBlockMap() to take a parent, and use the BlockMap's
ActionManager to enfore that block map growth is a suspended
operation.

.../recoveryJournal.h
.../recoveryJournal.c
Modified freeRecoveryJournal() to properly handle the case of freeing
a VDO which was suspended but not saved.
Modified resumeRecoveryJournal() to take a parent, to yield a read-only
error if the VDO is read-only, and to yield an invalid state error if
the journal is not quiescent.

.../packer.h
.../packer.c
Modified resumePacker() to take a parent.

.../slab.h
.../slab.c
Added an 'isNew' parameter to makeSlab() to indicate that the slab is
being created as part of a grow physical and not at startup. New
slabs must be quiescent (they start in the new ADMIN_STATE_NEW
state), and makeSlab() will allocate RefCounts for them.

.../superBlock.c
If a growth operation fails, the VDO is put into read-only mode, but
because the VDO has not yet been resumed, it is not allowed to
actually enter read-only mode. The notifications and such are done
during the resume operation. However, if the growth operation failed
because it couldn't update the super block with the new size, it
would be poor if the subsequent attempt to save the super block when
entering read-only mode during the resume were to succeed. This would
result both in an unnecessary read-only rebuild, and the confusion of
missing writes if the VDO is restarted read-only, which would then
reappear after a read-only rebuild. To solve this issue, an
'unwriteable' field was added to SuperBlock. This field is set if a
super block write fails, and when set, no subsequent super block
writes will be attempted.

.../slabDepot.h
.../slabDepot.c
Removed the 'reverting' parameter from udpateSlabDepotSize() since it
now only has one caller which is never reverting.
Modified resumeSlabDepot() to immediately finish the parent with a
read-only error if the VDO is read-only.

.../slabDepot.c
.../vdoLoad.c
Reflected changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>